+++ src/OkMD.jl
@@ -5,5 +5,6 @@ using Markdown
 include("collect.jl")
 export targetrange, targetsection
 
+using TOML
 include("parsefile.jl")
 end